requirements/dev.txt
alembic==1.14.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
cachelib==0.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
dnspython==2.7.0
email_validator==2.2.0
environs==11.2.1
factory_boy==3.3.1
Faker==33.1.0
flake8==7.1.1
flake8-blind-except==0.2.1
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-isort==6.1.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-DebugToolbar==0.16.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Static-Digest==0.4.1
Flask-WTF==1.2.2
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.2
mccabe==0.7.0
msgspec==0.19.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pep8-naming==0.14.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
six==1.17.0
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.36
supervisor==4.2.5
typing_extensions==4.12.2
urllib3==2.3.0
waitress==3.0.2
WebOb==1.8.9
WebTest==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
zope.event==5.0
zope.interface==7.2
